+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Added
 
 - Added deprecated functions and methods support.
-- Added support for references provider. The extension now knows where every reference to a symbol is (see [the documentation](https://code.visualstudio.com/api/language-extensions/programmatic-language-features#find-all-references-to-a-symbol) for more details.)
+- Added support for references provider. The extension now knows where every reference to a symbol is (see [the documentation](https://code.visualstudio.com/api/language-extensions/programmatic-language-features#find-all-references-to-a-symbol) for more details).
+- Added support for rename provider (see [the documentation](https://code.visualstudio.com/api/language-extensions/programmatic-language-features#rename-symbols) for more details).
 
 ### Fixed